Move CompositeScale to models/scales
