+++ bokehjs/src/lib/models/scales/index.ts
@@ -1,4 +1,5 @@
 export {CategoricalScale}         from "./categorical_scale"
+export {CompositeScale}           from "./composite_scale"
 export {ContinuousScale}          from "./continuous_scale"
 export {LinearScale}              from "./linear_scale"
 export {LinearInterpolationScale} from "./linear_interpolation_scale"
